fix: separated files structure
