--- entity.go
@@ -1,6 +1,0 @@
-type Pokemon struct {
-	ID     int    `json:ID`
-	Name   string `json:Name`
-	Weight int    `json:Weight`
-	Height int    `json:Height`
-}
